requirements.txt
streamlit~=1.55.0
nba_api>=1.11.4
requests~=2.31.0
curl_cffi>=0.7,<0.14
stripe~=14.4.0
aiohttp~=3.13.0
thefuzz~=0.22.0
streamlit-autorefresh~=1.0.1
flask~=3.1.0
python-dotenv~=1.0.0
pdfplumber~=0.11.0
pandas~=2.2.0
numpy~=1.26.0
scikit-learn~=1.4.0
xgboost~=2.0.0
catboost~=1.2.0
beautifulsoup4~=4.12.0
lxml~=5.1.0
fastapi~=0.110.0
uvicorn~=0.27.0
pyarrow~=15.0.0
joblib~=1.3.0
